EUREKA-777: Fix an issue when a role can't be found by name
