+++ src/folio_upm/integration/services/permission_service.py
@@ -1,7 +1,7 @@
 from folio_upm.dto.cls_support import SingletonMeta
 from folio_upm.integration.clients.permissions_client import PermissionsClient
 from folio_upm.utils import log_factory
-from folio_upm.utils.common_utils import CqlQueryUtils
+from folio_upm.utils.cql_query_utils import CqlQueryUtils
 from folio_upm.utils.loading_utils import PagedDataLoader, PartitionedDataLoader
 from folio_upm.utils.ordered_set import OrderedSet
 from folio_upm.utils.upm_env import Env
